requirements.txt
flask==2.3.3
openai==0.28.0
qrcode==7.4.2
Pillow==10.1.0
